pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "studybot-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "chromadb",
  "langchain-text-splitters",
  "playwright",
  "pydantic",
  "pyyaml",
  "anthropic",
  "google-genai",
  "zhipuai",
  "python-dotenv",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "httpx",
]

[tool.pytest.ini_options]
pythonpath = ["src/python"]
